Create a replacement for omnisearch
FUI provides two ways of doing something like that:
  - dropdown with search
  - search input

The first is only useful if we want to keep/select the result of the
search, as in our custom 'selectize' user input (as used in forms).
We can either keep one or multiple values.

The second is useful if you want search and click on something that
takes you somewhere, pretty much exactly what we do with our omnisearch
input.

This introduces item rendering based on class (ie: user, person,
competition, regulation), and a debounced request helper.